backend/requirements.txt
Django==3.2.12
djangorestframework==3.14.0
PyMySQL==1.0.2
django-cors-headers
gunicorn